Reading icon file

+++ src/PvBake.Lib/Models/Icon.cs
@@ -2,8 +2,8 @@ namespace PvBake.Lib.Models
 {
     public sealed class Icon : IFile
     {
-        public byte? Width { get; set; }
-        public byte? Height { get; set; }
+        public ushort? Width { get; set; }
+        public ushort? Height { get; set; }
         public byte[] Pixels { get; set; }
     }
 }
